requirements.txt
Django==5.1.2
flake8==7.1.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
pep8-naming==0.14.1
djangorestframework==3.15.2
